fix(backend): create back the entities folder and move entities to then
