+++ backend/my-way/src/main/java/br/ufrn/myway/Model/Entities/AbstractModel.java
@@ -1,9 +1,14 @@
-package br.ufrn.myway.Model;
-
-import com.fasterxml.jackson.annotation.JsonFormat;
-import jakarta.persistence.*;
+package br.ufrn.myway.Model.Entities;
 
 import java.time.LocalDate;
+
+import com.fasterxml.jackson.annotation.JsonFormat;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.MappedSuperclass;
 
 @MappedSuperclass
 public abstract class AbstractModel {
