+++ backend/my-way/src/main/java/br/ufrn/myway/Model/Entities/Goal.java
@@ -1,4 +1,4 @@
-package br.ufrn.myway.Model;
+package br.ufrn.myway.Model.Entities;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.JoinColumn;
