+++ backend/my-way/src/main/java/br/ufrn/myway/Model/Entities/User.java
@@ -1,4 +1,4 @@
-package br.ufrn.myway.Model;
+package br.ufrn.myway.Model.Entities;
 
 import br.ufrn.myway.Model.Enums.Roles;
 import jakarta.persistence.*;
