+++ backend/my-way/src/main/java/br/ufrn/myway/Repository/AbstractRepository.java
@@ -3,7 +3,7 @@ package br.ufrn.myway.Repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.NoRepositoryBean;
 
-import br.ufrn.myway.Model.AbstractModel;
+import br.ufrn.myway.Model.Entities.AbstractModel;
 
 @NoRepositoryBean
 public interface AbstractRepository<M extends AbstractModel> extends JpaRepository<M, Long> {}
