+++ backend/my-way/src/main/java/br/ufrn/myway/Repository/UserRepository.java
@@ -2,7 +2,7 @@ package br.ufrn.myway.Repository;
 
 import org.springframework.stereotype.Repository;
 
-import br.ufrn.myway.Model.User;
+import br.ufrn.myway.Model.Entities.User;
 
 import java.util.Optional;
