+++ backend/my-way/src/main/java/br/ufrn/myway/Service/AuthenticationService.java
@@ -6,7 +6,7 @@ import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 
-import br.ufrn.myway.Model.User;
+import br.ufrn.myway.Model.Entities.User;
 import br.ufrn.myway.Repository.UserRepository;
 
 @Service
