+++ backend/my-way/src/main/java/br/ufrn/myway/Service/PersonService.java
@@ -1,12 +1,12 @@
 package br.ufrn.myway.Service;
  
-import br.ufrn.myway.Model.Person;
-import br.ufrn.myway.Repository.PersonRepository; 
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.stereotype.Service;
+import java.util.List;
  
-import java.util.List; 
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+import br.ufrn.myway.Model.Entities.Person; 
+import br.ufrn.myway.Repository.PersonRepository;
 
 @Service
 public class PersonService {
