+++ backend/my-way/src/main/java/br/ufrn/myway/Service/UserService.java
@@ -1,7 +1,7 @@
 package br.ufrn.myway.Service;
 
+import br.ufrn.myway.Model.Entities.User;
 import br.ufrn.myway.Model.Enums.ErrorMessageUtils;
-import br.ufrn.myway.Model.User;
 import br.ufrn.myway.Repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.crypto.password.PasswordEncoder;
